The tools stop being part of the conversation loop
